Remove tests for removed directives and add tests for new directive
